Form remembers its prev state and functionality for relay on off controll
made the page so that form in live data and analytics now retains its previous state after a refresh or a new submit.
Also added new column to trLiveData that keeps track of whether the Relay should turn on or off.
This data is accesed through the TransformerData#isOn() method. The on off feature is user controlled. This is achieved by
a checkbox that every row in live data have that turns the relay of corresponding module id on or off.